Reduce run trace detail
Since we plan to do 100k iteration runs, with 10 runs each,
the recorded data will get very large.
Because we do not need the fine grained information
we just record every 100th iteration.

Because the start of the runs usually has the largest amount of change,
we record every tenth data point for the first 1000 iterations.

For 100k - 10 runs this will give us 10900 data points for each
algorithm and instance - a reduction by factor ~100.